ViewController -> VC로 파일명 변경, StartVC 간격수정, gauthBtnTapped()함수 호출 에러 수정
